refactor: move config/appConfig.ts to common
