+++ packages/staking-dashboard/src/App/App.tsx
@@ -4,7 +4,7 @@ import Intercom from 'react-intercom'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 import AppError from 'shared-components/AppError'
 import Web3ReactExporter from 'shared-components/Web3ReactExporter'
-import appConfig from 'src/config/appConfig'
+import appConfig from 'bzx-common/src/config/appConfig'
 import AppVM from './AppVM'
 import Footer from './Footer'
 import Header from './Header'
