+++ packages/staking-dashboard/src/domain/Web3ConnectionFactory.ts
@@ -3,7 +3,7 @@ import { Web3Wrapper } from '@0x/web3-wrapper'
 import { AbstractConnector } from '@web3-react/abstract-connector'
 import { ConnectorUpdate } from '@web3-react/types'
 import configProviders from '../config/providers.json'
-import appConfig from 'src/config/appConfig'
+import appConfig from 'bzx-common/src/config/appConfig'
 import ProviderType from 'bzx-common/src/domain/ProviderType'
 
 export default class Web3ConnectionFactory {
