+++ packages/staking-dashboard/src/index.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import TagManager from 'react-gtm-module'
 import App from './App'
-import appConfig from './config/appConfig'
+import appConfig from 'bzx-common/src/config/appConfig'
 import configProviders from './config/providers.json'
 import stakingProvider from './services/StakingProvider'
 import RootStore from './stores/RootStore'
